+++ modules/security/src/main/resources/sql/security/user/delete-roles.sql
@@ -1,0 +1,1 @@
+DELETE FROM security.user_roles WHERE user_id = :userId