+++ modules/security/src/main/resources/sql/security/user/insert-role.sql
@@ -1,0 +1,1 @@
+INSERT INTO security.user_roles (user_id, role) VALUES (:userId, :role)