+++ modules/security/src/main/resources/sql/security/user/lock-administrator-state.sql
@@ -1,0 +1,1 @@
+LOCK TABLE security.users, security.user_roles IN SHARE ROW EXCLUSIVE MODE