+++ modules/security/src/main/resources/sql/security/user/update.sql
@@ -1,0 +1,3 @@
+UPDATE security.users
+   SET username = :username, enabled = :enabled, updated_at = :updatedAt
+ WHERE id = :id